+++ source/matcher/h2_strcmp.hpp
@@ -1,42 +1,42 @@
 struct h2_matches_regex : h2_matches {
    const h2_string e;
    explicit h2_matches_regex(const h2_string& e_) : e(e_) {}
-   h2_fail* matches(const h2_string& a, int n, h2_mc c) const;
+   h2_fail* matches(const h2_string& a, size_t n, h2_mc c) const;
    virtual h2_line expection(h2_mc c) const override;
 };
 
 struct h2_matches_wildcard : h2_matches {
    const h2_string e;
    explicit h2_matches_wildcard(const h2_string& e_) : e(e_) {}
-   h2_fail* matches(const h2_string& a, int n, h2_mc c) const;
+   h2_fail* matches(const h2_string& a, size_t n, h2_mc c) const;
    virtual h2_line expection(h2_mc c) const override;
 };
 
 struct h2_matches_strcmp : h2_matches {
    const h2_string e;
    explicit h2_matches_strcmp(const h2_string& e_) : e(e_) {}
-   h2_fail* matches(const h2_string& a, int n, h2_mc c) const;
+   h2_fail* matches(const h2_string& a, size_t n, h2_mc c) const;
    virtual h2_line expection(h2_mc c) const override;
 };
 
 struct h2_matches_substr : h2_matches {
    const h2_string substring;
    explicit h2_matches_substr(const h2_string& substring_) : substring(substring_) {}
-   h2_fail* matches(const h2_string& a, int n, h2_mc c) const;
+   h2_fail* matches(const h2_string& a, size_t n, h2_mc c) const;
    virtual h2_line expection(h2_mc c) const override;
 };
 
 struct h2_matches_startswith : h2_matches {
    const h2_string prefix_string;
    explicit h2_matches_startswith(const h2_string& prefix_string_) : prefix_string(prefix_string_) {}
-   h2_fail* matches(const h2_string& a, int n, h2_mc c) const;
+   h2_fail* matches(const h2_string& a, size_t n, h2_mc c) const;
    virtual h2_line expection(h2_mc c) const override;
 };
 
 struct h2_matches_endswith : h2_matches {
    const h2_string suffix_string;
    explicit h2_matches_endswith(const h2_string& suffix_string_) : suffix_string(suffix_string_) {}
-   h2_fail* matches(const h2_string& a, int n, h2_mc c) const;
+   h2_fail* matches(const h2_string& a, size_t n, h2_mc c) const;
    virtual h2_line expection(h2_mc c) const override;
 };
 
@@ -44,7 +44,7 @@ struct h2_matches_json : h2_matches {
    const h2_string e;
    const h2_string selector;
    explicit h2_matches_json(const h2_string& e_, const h2_string& selector_) : e(e_), selector(selector_) {}
-   h2_fail* matches(const h2_string& a, int, h2_mc c) const;
+   h2_fail* matches(const h2_string& a, size_t, h2_mc c) const;
    virtual h2_line expection(h2_mc c) const override;
 };
 
@@ -53,7 +53,7 @@ struct h2_caseless_matches : h2_matches {
    explicit h2_caseless_matches(const h2_matcher<h2_string>& matcher_) : m(matcher_) {}
 
    template <typename A>
-   h2_fail* matches(const A& a, int n, h2_mc c) const { return m.matches(a, n, c.update_caseless(true)); }
+   h2_fail* matches(const A& a, size_t n, h2_mc c) const { return m.matches(a, n, c.update_caseless(true)); }
    virtual h2_line expection(h2_mc c) const override { return m.expection(c.update_caseless(true)); }
 };
 
@@ -62,7 +62,7 @@ struct h2_spaceless_matches : h2_matches {
    explicit h2_spaceless_matches(const h2_matcher<h2_string>& matcher_) : m(matcher_) {}
 
    template <typename A>
-   h2_fail* matches(const A& a, int n, h2_mc c) const { return m.matches(a, n, c.update_spaceless(true)); }
+   h2_fail* matches(const A& a, size_t n, h2_mc c) const { return m.matches(a, n, c.update_spaceless(true)); }
    virtual h2_line expection(h2_mc c) const override { return m.expection(c.update_spaceless(true)); }
 };
